[refactor] 폴더위치 변경과 그로인한 수정사항들
