+++ src/main/java/com/haejwo/tripcometrue/domain/triprecord/dto/response/ModelAttribute/TripRecordListRequestAttribute.java
@@ -1,4 +1,4 @@
-package com.haejwo.tripcometrue.domain.triprecord.dto;
+package com.haejwo.tripcometrue.domain.triprecord.dto.response.ModelAttribute;
 
 public record TripRecordListRequestAttribute(
     String hashtag,
