+++ src/main/java/com/haejwo/tripcometrue/domain/triprecord/repository/triprecord/TripRecordCustomRepository.java
@@ -1,6 +1,6 @@
 package com.haejwo.tripcometrue.domain.triprecord.repository.triprecord;
 
-import com.haejwo.tripcometrue.domain.triprecord.dto.TripRecordListRequestAttribute;
+import com.haejwo.tripcometrue.domain.triprecord.dto.response.ModelAttribute.TripRecordListRequestAttribute;
 import com.haejwo.tripcometrue.domain.triprecord.entity.TripRecord;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
